refactor: landing hero and contact
